+++ Libplanet/Hashcash.cs
@@ -21,12 +21,12 @@ namespace Libplanet
         /// should not vary for different <paramref name="nonce"/>s.</para>
         /// </summary>
         /// <param name="nonce">An arbitrary nonce for an attempt, provided
-        /// by <see cref="Hashcash.Answer(Stamp, int)"/> method.</param>
+        /// by <see cref="Hashcash.Answer(Stamp, long)"/> method.</param>
         /// <returns>A <see cref="byte"/> array determined from the given
         /// <paramref name="nonce"/>.  It should return consistently
         /// an equivalent array for equivalent <paramref name="nonce"/>
         /// values.</returns>
-        /// <seealso cref="Hashcash.Answer(Stamp, int)"/>
+        /// <seealso cref="Hashcash.Answer(Stamp, long)"/>
         /// <seealso cref="Nonce"/>
         public delegate byte[] Stamp(Nonce nonce);
 
@@ -45,7 +45,7 @@ namespace Libplanet
         /// <returns>A <see cref="Nonce"/> value which satisfies the given
         /// <paramref name="difficulty"/>.</returns>
         /// <seealso cref="Stamp"/>
-        public static Nonce Answer(Stamp stamp, int difficulty)
+        public static Nonce Answer(Stamp stamp, long difficulty)
         {
             var nonceBytes = new byte[10];
             var random = new Random();
